refactor: merge internal/pkg/utils and internal/backend/utils into internal/utils
