+++ backend/utils/image_validation.go
