+++ backend/utils/types.go
@@ -1,0 +1,5 @@
+package utils
+
+type UserID string
+
+var User_id UserID = "userID"
